% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gingeR.R
\name{gingeR}
\alias{gingeR}
\title{gingeR}
\usage{
gingeR(name, format_file, auth_type, endpoint = NULL, tmpDir = "~\\\\")
}
\arguments{
\item{name}{name of the service}

\item{format_file}{format of the files providedd by the service}

\item{auth_type}{authentication type: "none", "basic", ...}

\item{endpoint}{endpoint in case auth_type is oauth1 or oauth2}

\item{tmpDir}{path of temporary folder}
}
\description{
gingeR
}
\examples{
\dontrun{
gingeR()
}
}
\keyword{gingeR}

